+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - `curry()` function
+- `k()` (Kestrel) function
 
 ### Removed